api/requirements/app.txt
django
djangorestframework
django-cors-headers
gunicorn
django-utils
celery

psycopg2
psycopg2-binary

google-cloud-storage

inflection

redis
httpx
logdna
qrcode